+++ menu/src/main/java/kitchenpos/menu/ui/MenuRestController.java
@@ -1,7 +1,8 @@
-package kitchenpos.menu.application;
+package kitchenpos.menu.ui;
 
 import java.net.URI;
 import java.util.List;
+import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.dto.request.MenuCreateRequest;
 import kitchenpos.menu.dto.response.MenuResponse;
 import org.springframework.http.ResponseEntity;
